--- a/src/components/Grid/Grid.vue
+++ b/src/components/Grid/Grid.vue
@@ -193,6 +193,19 @@ export default {
       get() { return this.$store.state.project.view.max_y; },
       set(val) { this.$store.dispatch('project/setViewMaxY', { max_y: val }); },
     },
+    dims: {
+      get() { 
+        return {
+          min_x: this.$store.state.project.view.min_x,
+          max_x: this.$store.state.project.view.max_x,
+          min_y: this.$store.state.project.view.min_y,
+          max_y: this.$store.state.project.view.max_y,
+        };
+      },
+      set(val) {
+        this.$store.dispatch('project/setDimensions', val);
+      },
+    },
     // previous story
     previousStory() {
       const currentStoryNumber = this.$store.state.models.stories.findIndex(s => s === this.currentStory);