--- a/src/store/modules/geometry/actions/componentPreservationSociety.js
+++ b/src/store/modules/geometry/actions/componentPreservationSociety.js
@@ -127,7 +127,6 @@ function replaceComponents(
       }, { root: true });
     });
   });
-  console.log(performance.now() - t0);
 }
 
 export function withPreservedComponents(...args) {