--- a/src/store/modules/geometry/helpers.js
+++ b/src/store/modules/geometry/helpers.js
@@ -283,12 +283,16 @@ const helpers = {
     /*
      * return the set of saved vertices directly on an edge, not including edge endpoints
      */
-  splittingVerticesForEdgeId(edge_id, geometry, spacing) {
-    const edge = geometry.edges.find(e => e.id === edge_id),
-      edgeV1 = this.vertexForId(edge.v1, geometry),
+  splittingVerticesForEdgeId(edge, geometry, spacing, vertices) {
+    const edgeV1 = this.vertexForId(edge.v1, geometry),
       edgeV2 = this.vertexForId(edge.v2, geometry);
       // look up all vertices touching the edge, ignoring the edge's endpoints
-    const verticesToSplit = geometry.vertices.filter((vertex) => {
+
+    if (!vertices) {
+      vertices = geometry.vertices;
+    }
+    
+    return vertices.filter((vertex) => {
       const
         vertexIsEndpointById = edge.v1 === vertex.id || edge.v2 === vertex.id,
         vertexIsLeftEndpointByValue = edgeV1.x === vertex.x && edgeV1.y === vertex.y,
@@ -307,7 +311,6 @@ const helpers = {
       const shouldSplit = distBetween <= spacing / 20;
       return shouldSplit;
     });
-    return verticesToSplit;
   },
 
   projectionOfPointToLine,