… package.json
… src/store/modules/models/actions.js
… src/store/modules/models/mutations.js
… src/store/modules/project/actions.js
… src/store/timetravel.js
… test/e2e/nightwatch.conf.js
… test/e2e/specs/test.js
… test/perf/.eslintrc
… test/perf/datasets/largePlan.json
… test/perf/datasets/largeShatteredPlan.json
… test/perf/datasets/smallPlan.json
… test/perf/index.js
… test/perf/karma.conf.js
… test/perf/specs/add.spec.js
… test/perf/specs/helper.js
… test/perf/specs/zoom.spec.js
… test/unit/specs/geometry/helpers.spec.js