@benfen 2021-07-15
There's still a lot of stuff left to do and the perf tests could use some love, but things are stable for now and I'd like to get it merged in so it's not floating out here for too long.